Convert SymbolsSpec.js -> Symbols.spec.tsx
