+++ test/index.js
@@ -37,5 +37,4 @@ require('./specs/shape/DotSpec');
 require('./specs/shape/RectangleSpec');
 require('./specs/shape/SectorSpec');
 require('./specs/shape/PolygonSpec');
-require('./specs/shape/SymbolsSpec');
 require('./specs/shape/TrapezoidSpec');
